fix: fix hotel suggestion error, fix zustand store data error, add search bar inside search result page, separated search bar as a component
